% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_ApplicationDbContext.R
\docType{data}
\name{ApplicationDbContext}
\alias{ApplicationDbContext}
\title{Base class for the application database context}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
ApplicationDbContext
}
\value{
Object of type R6Class with methods for communication with the database.
}
\description{
Base class for the application database context
}
\examples{
\dontrun{
mongoConnectionString = "mongodb+srv://admin:example"

MongoDbContext <- ApplicationDbContext$new(
  connectionString = mongoConnectionString,
  database = "ExampleDatabase",
  collection = "ExampleCollection"
)
}

}
\keyword{datasets}
