% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_SchemaValidator.R
\name{CompaniesSchemaValidator}
\alias{CompaniesSchemaValidator}
\title{Creates a JSON schema validator for Companies data}
\usage{
CompaniesSchemaValidator()
}
\value{
jsonvalidator object
}
\description{
Creates a JSON schema validator for Companies data
}
\examples{
CompaniesSchemaValidator()
}
