% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_CompaniesService.R
\docType{data}
\name{CompaniesService}
\alias{CompaniesService}
\title{Service for interacting with Companies data.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompaniesService
}
\arguments{
\item{DbContext}{ApplicationDbContext. Database context for connecting to the
Companies data. This should be an instance of the ApplicationDbContext class.}
}
\description{
Service for interacting with Companies data.
}
\section{Methods}{

\code{$new()} Initialize a new Company Service instance.

\code{$getCompaniesCount()} Returns the number of Companies.

\code{$getCompanies()} Returns a data frame of all Companies.

\code{$getCompany(companyName)} Returns a data frame for the required Company name.

\code{$getOldestCompanies(limit)} Returns a data frame of the oldest companies, for the required limit.
}

\examples{
\dontrun{
mongoConnectionString <- "mongodb+srv://admin:example"

MongoDbContext <- ApplicationDbContext$new(
  connectionString = mongoConnectionString,
  database = "ExampleDatabase",
  collection = "ExampleCollection"
)

testCompaniesService <- CompaniesService$new(DbContext = MongoDbContext)
}

}
\keyword{datasets}
