% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_DivideTwoNumbers.R
\name{divideTwoNumbers}
\alias{divideTwoNumbers}
\title{Divides two numbers}
\usage{
divideTwoNumbers(numerator, denominator)
}
\arguments{
\item{numerator}{Numeric. Numerator to use in division.}

\item{denominator}{Numeric. Denominator to use in division.}
}
\value{
Division result. Numeric.
}
\description{
Divides two numbers
}
\examples{
divideTwoNumbers(numerator = 4, denominator = 2)
}
