% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers_Logger.R
\name{InitialiseLogger}
\alias{InitialiseLogger}
\title{Initialises a new logger}
\usage{
InitialiseLogger(
  logLevel = logger::DEBUG,
  logMessageFormat = "[{level}] [{time}] [{namespace}] [{fn}]: {msg}"
)
}
\arguments{
\item{logLevel}{The required log level threshold to capture in the logger.}

\item{logMessageFormat}{The required log message format.
Defaults to [{level}] [{time}] [{namespace}] [{fn}]: {msg}. See logger::layout_glue_generator for options.}
}
\value{
NULL
#' @examples
\dontrun{
# Using  default settings:
InitialiseLogger()

# Using another log message format:
InitialiseLogger(logLevel = logger::DEBUG, logMessageFormat = '{level}---{time}--- msg')
}
}
\description{
Initialises a new logger
}
