% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_ApplicationDbContext.R
\docType{data}
\name{ApplicationDbContext}
\alias{ApplicationDbContext}
\title{Base class for the application database context}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationDbContext
}
\arguments{
\item{connectionString}{String. Connection string to use.}

\item{database}{String. Database to connect to.}

\item{collection}{String. Database collection to connect to.}

\item{verbose}{Boolean. Option to emit more connection messages. Defaults to TRUE.}
}
\value{
Object of type \code{\link{R6Class}} with methods for communication with the database.
}
\description{
Base class for the application database context
}
\section{Fields}{

\describe{
\item{\code{DbConnection}}{MongoDB connection object from the \code{\link{mongo}} function.}
}}

\examples{
\dontrun{
mongoConnectionString = "mongodb+srv://admin:example"

MongoDbContext <- ApplicationDbContext$new(
  connectionString = mongoConnectionString,
  database = "ExampleDatabase",
  collection = "ExampleCollection"
)
}

}
\keyword{datasets}
