% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_ApplicationDbContext.R
\name{ApplicationDbContext}
\alias{ApplicationDbContext}
\title{Base class for the application database context}
\value{
Object of type R6Class with methods for communication with the database.
}
\description{
Base class for the application database context

Base class for the application database context
}
\examples{
\dontrun{
mongoConnectionString = "mongodb+srv://admin:example"

MongoDbContext <- ApplicationDbContext$new(
  connectionString = mongoConnectionString,
  database = "ExampleDatabase",
  collection = "ExampleCollection"
)
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{DbConnection}}{Database Connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApplicationDbContext$new()}}
\item \href{#method-clone}{\code{ApplicationDbContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ApplicationDbContext object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationDbContext$new(
  connectionString,
  database,
  collection,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionString}}{Connection string.}

\item{\code{database}}{Database name.}

\item{\code{collection}}{Collection name.}

\item{\code{verbose}}{Verbose setting. Defaults to TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ApplicationDbContext object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationDbContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
