% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_CompaniesService.R
\docType{data}
\name{CompaniesService}
\alias{CompaniesService}
\title{Service for interacting with Companies data.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompaniesService
}
\value{
Object of type R6Class with methods for communication with Companies data.
}
\description{
Service for interacting with Companies data.
}
\section{Methods}{

new() Initialize a new Company Service instance.

getCompaniesCount() Returns the number of Companies.

getCompanies() Returns a data frame of all Companies.

getCompany(companyName) Returns a data frame for the required Company name.

getOldestCompanies(limit) Returns a data frame of the oldest companies, for the required limit.

getNumberOfCompaniesFoundedPerYear() Returns the number of companies founded per year.
}

\examples{
\dontrun{
mongoConnectionString <- "mongodb+srv://admin:example"

MongoDbContext <- ApplicationDbContext$new(
  connectionString = mongoConnectionString,
  database = "ExampleDatabase",
  collection = "ExampleCollection"
)

testCompaniesService <- CompaniesService$new(dbContext = MongoDbContext)
}

}
\keyword{datasets}
