% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_CompaniesService.R
\name{CompaniesService}
\alias{CompaniesService}
\title{Service for interacting with Companies data.}
\value{
Object of type R6Class with methods for communication with Companies data.
}
\description{
Service for interacting with Companies data.

Service for interacting with Companies data.
}
\examples{
\dontrun{
mongoConnectionString <- "mongodb+srv://admin:example"

MongoDbContext <- ApplicationDbContext$new(
  connectionString = mongoConnectionString,
  database = "ExampleDatabase",
  collection = "ExampleCollection"
)

testCompaniesService <- CompaniesService$new(dbContext = MongoDbContext)
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CompaniesService$new()}}
\item \href{#method-getCompaniesCount}{\code{CompaniesService$getCompaniesCount()}}
\item \href{#method-getCompanies}{\code{CompaniesService$getCompanies()}}
\item \href{#method-getCompany}{\code{CompaniesService$getCompany()}}
\item \href{#method-getCompanyId}{\code{CompaniesService$getCompanyId()}}
\item \href{#method-getOldestCompanies}{\code{CompaniesService$getOldestCompanies()}}
\item \href{#method-getNumberOfCompaniesFoundedPerYear}{\code{CompaniesService$getNumberOfCompaniesFoundedPerYear()}}
\item \href{#method-createCompany}{\code{CompaniesService$createCompany()}}
\item \href{#method-editCompany}{\code{CompaniesService$editCompany()}}
\item \href{#method-deleteCompany}{\code{CompaniesService$deleteCompany()}}
\item \href{#method-clone}{\code{CompaniesService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a new CompaniesService object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$new(dbContext, schemaValidator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbContext}}{ApplicationDbContext object.}

\item{\code{schemaValidator}}{SchemaValidator object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CompaniesService object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompaniesCount"></a>}}
\subsection{Method \code{getCompaniesCount()}}{
Gets the number of Companies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$getCompaniesCount()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompanies"></a>}}
\subsection{Method \code{getCompanies()}}{
Returns a data frame of all Companies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$getCompanies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompany"></a>}}
\subsection{Method \code{getCompany()}}{
Returns a data frame for the required Company name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$getCompany(companyName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companyName}}{Company name string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompanyId"></a>}}
\subsection{Method \code{getCompanyId()}}{
Returns the Id of the required Company name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$getCompanyId(companyName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companyName}}{Company name string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOldestCompanies"></a>}}
\subsection{Method \code{getOldestCompanies()}}{
Returns a data frame of the oldest Companies, for the required limit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$getOldestCompanies(limit = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Number of Companies to limit the results to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNumberOfCompaniesFoundedPerYear"></a>}}
\subsection{Method \code{getNumberOfCompaniesFoundedPerYear()}}{
Returns the number of Companies founded per year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$getNumberOfCompaniesFoundedPerYear()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createCompany"></a>}}
\subsection{Method \code{createCompany()}}{
Creates a new Company record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$createCompany(companyDetails)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companyDetails}}{JSON string with new Company details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-editCompany"></a>}}
\subsection{Method \code{editCompany()}}{
Edits an existing Company record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$editCompany(companyName, newCompanyDetails)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companyName}}{Company name string.}

\item{\code{newCompanyDetails}}{JSON string with new Company details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteCompany"></a>}}
\subsection{Method \code{deleteCompany()}}{
Deletes an existing Company record.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$deleteCompany(companyName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companyName}}{Company name string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
